requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
certifi==2024.07.04
charset-normalizer==3.3.2
discord.py==2.4.0
diskcache==5.6.3
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
llama_cpp_python==0.2.79
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.0.0
packaging==24.1
PyYAML==6.0.1
requests==2.32.3
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
